refactor: 비디오 관련 리팩토링 및 subscribers 화면 삭제
