+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "mocha.subdirectory": "src/components/class/Video"
+}
