+++ src/components/class/ChatFrame.jsx
@@ -11,9 +11,9 @@ import {
   messageEvent,
   leaveEvent,
 } from "../../modules/redux/cookingClass";
-import Chats from "./Chat/Chats";
-import CreateChat from "./Chat/CreateChat";
-import ChatHeader from "./Chat/ChatHeader";
+import Chats from "./chat/Chats";
+import CreateChat from "./chat/CreateChat";
+import ChatHeader from "./chat/ChatHeader";
 
 // TODO change v4 -> v5
 const ChatFrame = (props) => {
