+++ src/components/class/VideoFrame.jsx
@@ -1,11 +1,11 @@
-import VideoHeader from "./VideoHeader";
-import Video from "./Video";
+import VideoHeader from "./video/VideoHeader";
+import OpenViduVideo from "./video/OpenViduVideo";
 
 const VideoFrame = () => {
   return (
     <>
       <VideoHeader />
-      <Video />
+      <OpenViduVideo />
     </>
   );
 };
